Refactor to project, divides project into operations, tensors, overloads, tests and gpu
