+++ gpu/cudaOps.cu
