+++ main.cpp
@@ -1,4 +1,4 @@
-#include "tfclone.h"
+#include "buildTensorflow.h"
 
 void oldSigmoidTest() {
