+++ makefile
@@ -1,13 +1,13 @@
 all: main.cpp
-	g++ -std=c++11 main.cpp -o build/main.o
+	g++ -std=c++11 -I . main.cpp -o build/main.o
 	./build/main.o
 debug: main.cpp
-	g++ -g -std=c++11 main.cpp -o build/main.o
+	g++ -g -std=c++11 -I . main.cpp -o build/main.o
 	./build/main.o
-test: test.cpp
-	g++ -std=c++11 test.cpp -lgtest -lgtest_main -pthread -o build/test.o
+test: tests/test.cpp
+	g++ -std=c++11 -I . tests/test.cpp -lgtest -lgtest_main -pthread -o build/test.o
 	./build/test.o
-gpuTest: cudaOps.cu
-	nvcc -std=c++14 cudaOps.cu -o build/cuda.o
+gpuTest: gpu/cudaOps.cu
+	nvcc -std=c++14 -I . gpu/cudaOps.cu -o build/cuda.o
 	./build/cuda.o
