+++ operations/operations_Impl.h
@@ -1,0 +1,15 @@
+/*
+This file includes all the operator implementations
+*/
+
+#ifndef __OP_IMPL_INCLUDED__   
+#define __OP_IMPL_INCLUDED__  
+
+#include "operations/multiplyOperation_Impl.h"
+#include "operations/addOperation_Impl.h"
+#include "operations/divideOperation_Impl.h"
+#include "operations/dotOperation_Impl.h"
+#include "operations/exponentOperation_Impl.h"
+
+#endif
+
