+++ tests/test.cpp
@@ -1,6 +1,6 @@
 #include<iostream>
 #include <gtest/gtest.h>
-#include "matrix.h"
+#include "tensor/matrix.h"
 
 using namespace std;
