--- tfclone.h
@@ -1,3 +1,0 @@
-#include "tensor.h" 
-#include "operations_Impl.h"
-
